host/requirements.txt
pillow
google-generativeai
websockets